requirements.txt
streamlit>=1.30.0
openai>=1.10.0
pandas>=2.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
pytesseract>=0.3.10
plotly>=5.18.0
